feat: add example for columns
