+++ dev-example/data/columnListBlocks.json
@@ -1,0 +1,168 @@
+{
+    "object": "list",
+    "results": [{
+    "object": "block",
+    "id": "dummy-id-1",
+    "parent": {
+      "type": "page_id", 
+      "page_id": "dummy-page-id-1"
+    },
+    "created_time": "2024-12-05T10:34:00.000Z",
+    "last_edited_time": "2024-12-05T10:34:00.000Z",
+    "created_by": {
+      "object": "user",
+      "id": "dummy-user-id-1"
+    },
+    "last_edited_by": {
+      "object": "user",
+      "id": "dummy-user-id-1" 
+    },
+    "has_children": true,
+    "archived": false,
+    "in_trash": false,
+    "type": "column_list",
+    "column_list": {
+      "children": [
+        {
+          "object": "block",
+          "id": "dummy-id-2",
+          "parent": {
+            "type": "block_id",
+            "block_id": "dummy-id-1"
+          },
+          "created_time": "2024-12-05T10:34:00.000Z",
+          "last_edited_time": "2024-12-05T10:34:00.000Z",
+          "created_by": {
+            "object": "user",
+            "id": "dummy-user-id-1"
+          },
+          "last_edited_by": {
+            "object": "user",
+            "id": "dummy-user-id-1"
+          },
+          "has_children": true,
+          "archived": false,
+          "in_trash": false,
+          "type": "column",
+          "column": {
+            "children": [
+              {
+                "object": "block",
+                "id": "dummy-id-3",
+                "parent": {
+                  "type": "block_id",
+                  "block_id": "dummy-id-2"
+                },
+                "created_time": "2024-12-05T10:34:00.000Z",
+                "last_edited_time": "2024-12-05T13:23:00.000Z",
+                "created_by": {
+                  "object": "user",
+                  "id": "dummy-user-id-1"
+                },
+                "last_edited_by": {
+                  "object": "user",
+                  "id": "dummy-user-id-1"
+                },
+                "has_children": false,
+                "archived": false,
+                "in_trash": false,
+                "type": "paragraph",
+                "paragraph": {
+                  "rich_text": [
+                    {
+                      "type": "text",
+                      "text": {
+                        "content": "LeftCol",
+                        "link": null
+                      },
+                      "annotations": {
+                        "bold": false,
+                        "italic": false,
+                        "strikethrough": false,
+                        "underline": false,
+                        "code": false,
+                        "color": "default"
+                      },
+                      "plain_text": "LeftCol",
+                      "href": null
+                    }
+                  ],
+                  "color": "default"
+                }
+              }
+            ]
+          }
+        },
+        {
+          "object": "block",
+          "id": "dummy-id-4",
+          "parent": {
+            "type": "block_id",
+            "block_id": "dummy-id-1"
+          },
+          "created_time": "2024-12-05T10:34:00.000Z",
+          "last_edited_time": "2024-12-05T10:34:00.000Z",
+          "created_by": {
+            "object": "user",
+            "id": "dummy-user-id-1"
+          },
+          "last_edited_by": {
+            "object": "user",
+            "id": "dummy-user-id-1"
+          },
+          "has_children": true,
+          "archived": false,
+          "in_trash": false,
+          "type": "column",
+          "column": {
+            "children": [
+              {
+                "object": "block",
+                "id": "dummy-id-5",
+                "parent": {
+                  "type": "block_id",
+                  "block_id": "dummy-id-4"
+                },
+                "created_time": "2024-12-05T10:34:00.000Z",
+                "last_edited_time": "2024-12-05T13:23:00.000Z",
+                "created_by": {
+                  "object": "user",
+                  "id": "dummy-user-id-1"
+                },
+                "last_edited_by": {
+                  "object": "user",
+                  "id": "dummy-user-id-1"
+                },
+                "has_children": false,
+                "archived": false,
+                "in_trash": false,
+                "type": "paragraph",
+                "paragraph": {
+                  "rich_text": [
+                    {
+                      "type": "text",
+                      "text": {
+                        "content": "RightCol",
+                        "link": null
+                      },
+                      "annotations": {
+                        "bold": false,
+                        "italic": false,
+                        "strikethrough": false,
+                        "underline": false,
+                        "code": false,
+                        "color": "default"
+                      },
+                      "plain_text": "RightCol",
+                      "href": null
+                    }
+                  ],
+                  "color": "default"
+                }
+              }
+            ]
+          }
+        }
+      ]
+    }
+  }]}
