+++ dev-example/pages/[id].js
@@ -1,7 +1,7 @@
 import React from 'react'
 import Head from 'next/head'
 import NextImg from 'next/image'
-import { getDatabase, getPage, getBlocks } from '../lib/notion'
+import { getDatabase, getPage, getBlocks, getBlocksWithChildren } from '../lib/notion'
 import Link from 'next/link'
 import { databaseId } from './blog.js'
 
@@ -27,6 +27,7 @@ export default function Post({ page, blocks }) {
   if (!page || !blocks) {
     return <div />
   }
+
 
   return (
     <>
@@ -69,29 +70,31 @@ export const getStaticPaths = async () => {
 export const getStaticProps = async (context) => {
   const { id } = context.params
   const page = await getPage(id)
-  const blocks = await getBlocks(id)
+  // const blocks = await getBlocks(id)
 
   // Retrieve block children for nested blocks (one level deep), for example toggle blocks
   // https://developers.notion.com/docs/working-with-page-content#reading-nested-blocks
-  const childBlocks = await Promise.all(
-    blocks
-      .filter((block) => block.has_children)
-      .map(async (block) => {
-        return {
-          id: block.id,
-          children: await getBlocks(block.id)
-        }
-      })
-  )
-  const blocksWithChildren = blocks.map((block) => {
-    // Add child blocks if the block should contain children but none exists
-    if (block.has_children && !block[block.type].children) {
-      block[block.type].children = childBlocks.find(
-        (x) => x.id === block.id
-      )?.children
-    }
-    return block
-  })
+  // const childBlocks = await Promise.all(
+  //   blocks
+  //     .filter((block) => block.has_children)
+  //     .map(async (block) => {
+  //       return {
+  //         id: block.id,
+  //         children: await getBlocks(block.id)
+  //       }
+  //     })
+  // )
+  // const blocksWithChildren = blocks.map((block) => {
+  //   // Add child blocks if the block should contain children but none exists
+  //   if (block.has_children && !block[block.type].children) {
+  //     block[block.type].children = childBlocks.find(
+  //       (x) => x.id === block.id
+  //     )?.children
+  //   }
+  //   return block
+  // })
+
+  const blocksWithChildren = await getBlocksWithChildren(id)
 
   return {
     props: {
