+++ dev-example/pages/index.js
@@ -1,9 +1,11 @@
 import React from 'react'
 import Link from 'next/link'
 import { Render } from '@9gustin/react-notion-render'
+// import "../../dist/index.css" // <-- this is the css file from the package
 
 import notionResponse from '../data/mockVideos.json'
 import title from '../data/title.json'
+import columnListResponse from '../data/columnListBlocks.json'
 
 export default function mockedPage() {
   return (
@@ -12,9 +14,10 @@ export default function mockedPage() {
         This page is mockuped with data/blocks.json, also you can view{' '}
         <Link href='/blog'>/blog</Link>
       </h3>
-      <Render blocks={[title.properties.Name]}/>
+      <Render blocks={[title.properties.Name]} />
       <article>
-        <Render blocks={notionResponse.results} classNames/>
+        <Render blocks={columnListResponse.results} classNames />
+        <Render blocks={notionResponse.results} classNames />
       </article>
     </div>
   )
